Move storage to modules
